# openapi/avedra-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AVEDRA API",
    "description": "API REST de AVEDRA — Asistencia, Convivencia y Evaluacion",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/contexto/grupos": {
      "get": {
        "tags": [
          "contexto"
        ],
        "summary": "Listar Grupos",
        "description": "Lista todos los grupos del catalogo (opcionalmente filtrados por grado).",
        "operationId": "listar_grupos_api_v1_contexto_grupos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "grado",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Grado"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GrupoResponse"
                  },
                  "title": "Response Listar Grupos Api V1 Contexto Grupos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contexto/asignaturas": {
      "get": {
        "tags": [
          "contexto"
        ],
        "summary": "Listar Asignaturas",
        "description": "Lista todas las asignaturas del catalogo (opcionalmente por area).",
        "operationId": "listar_asignaturas_api_v1_contexto_asignaturas_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "area_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Area Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AsignaturaResponse"
                  },
                  "title": "Response Listar Asignaturas Api V1 Contexto Asignaturas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contexto/periodos": {
      "get": {
        "tags": [
          "contexto"
        ],
        "summary": "Listar Periodos",
        "description": "Lista los periodos de un anio lectivo.",
        "operationId": "listar_periodos_api_v1_contexto_periodos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "anio_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Anio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PeriodoResponse"
                  },
                  "title": "Response Listar Periodos Api V1 Contexto Periodos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contexto/periodos/activo": {
      "get": {
        "tags": [
          "contexto"
        ],
        "summary": "Periodo Activo",
        "description": "Retorna el periodo activo del anio lectivo.",
        "operationId": "periodo_activo_api_v1_contexto_periodos_activo_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "anio_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Anio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeriodoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contexto/anio-activo": {
      "get": {
        "tags": [
          "contexto"
        ],
        "summary": "Anio Activo",
        "description": "Retorna la configuracion del anio lectivo activo.",
        "operationId": "anio_activo_api_v1_contexto_anio_activo_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/convivencia/observaciones": {
      "get": {
        "tags": [
          "convivencia"
        ],
        "summary": "Listar Observaciones",
        "description": "Lista observaciones de un estudiante (opcionalmente filtradas por periodo).",
        "operationId": "listar_observaciones_api_v1_convivencia_observaciones_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "estudiante_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Estudiante Id"
            }
          },
          {
            "name": "periodo_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Periodo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ObservacionResponse"
                  },
                  "title": "Response Listar Observaciones Api V1 Convivencia Observaciones Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "convivencia"
        ],
        "summary": "Registrar Observacion",
        "description": "Registra o actualiza una observacion narrativa.",
        "operationId": "registrar_observacion_api_v1_convivencia_observaciones_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NuevaObservacionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservacionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/convivencia/resumen-grupo": {
      "get": {
        "tags": [
          "convivencia"
        ],
        "summary": "Resumen Convivencia Grupo",
        "description": "Resumen de convivencia (notas, observaciones, alertas) de todos los estudiantes.",
        "operationId": "resumen_convivencia_grupo_api_v1_convivencia_resumen_grupo_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "grupo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          },
          {
            "name": "periodo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Periodo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumenConvivenciaResponse"
                  },
                  "title": "Response Resumen Convivencia Grupo Api V1 Convivencia Resumen Grupo Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/convivencia/notas-grupo": {
      "get": {
        "tags": [
          "convivencia"
        ],
        "summary": "Notas Grupo",
        "description": "Retorna las notas de comportamiento de todos los estudiantes del grupo.",
        "operationId": "notas_grupo_api_v1_convivencia_notas_grupo_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "grupo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          },
          {
            "name": "periodo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Periodo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotaComportamientoResponse"
                  },
                  "title": "Response Notas Grupo Api V1 Convivencia Notas Grupo Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/convivencia/vista-360/{estudiante_id}": {
      "get": {
        "tags": [
          "convivencia"
        ],
        "summary": "Vista 360",
        "description": "Perfil 360 de convivencia de un estudiante en un periodo.",
        "operationId": "vista_360_api_v1_convivencia_vista_360__estudiante_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "estudiante_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Estudiante Id"
            }
          },
          {
            "name": "periodo_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Periodo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AsignaturaResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "codigo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo"
          },
          "area_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Id"
          },
          "institucion_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institucion Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre"
        ],
        "title": "AsignaturaResponse"
      },
...
      "GrupoResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "grado": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grado"
          },
          "codigo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo"
          },
          "institucion_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institucion Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre"
        ],
        "title": "GrupoResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotaComportamientoResponse": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "estudiante_id": {
            "type": "integer",
            "title": "Estudiante Id"
          },
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "periodo_id": {
            "type": "integer",
            "title": "Periodo Id"
          },
          "valor": {
            "type": "number",
            "title": "Valor"
          },
          "observacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacion"
          }
        },
        "type": "object",
        "required": [
          "estudiante_id",
          "grupo_id",
          "periodo_id",
          "valor"
        ],
        "title": "NotaComportamientoResponse"
      },
      "NuevaObservacionRequest": {
        "properties": {
          "estudiante_id": {
            "type": "integer",
            "title": "Estudiante Id"
          },
          "asignacion_id": {
            "type": "integer",
            "title": "Asignacion Id"
          },
          "periodo_id": {
            "type": "integer",
            "title": "Periodo Id"
          },
          "texto": {
            "type": "string",
            "title": "Texto"
          },
          "categoria_id": {
            "type": "integer",
            "title": "Categoria Id"
          },
          "es_publica": {
            "type": "boolean",
            "title": "Es Publica",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "estudiante_id",
          "asignacion_id",
          "periodo_id",
          "texto",
          "categoria_id"
        ],
        "title": "NuevaObservacionRequest"
      },
...
      "ObservacionResponse": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "estudiante_id": {
            "type": "integer",
            "title": "Estudiante Id"
          },
          "asignacion_id": {
            "type": "integer",
            "title": "Asignacion Id"
          },
          "periodo_id": {
            "type": "integer",
            "title": "Periodo Id"
          },
          "texto": {
            "type": "string",
            "title": "Texto"
          },
          "categoria_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria Id"
          },
          "es_publica": {
            "type": "boolean",
            "title": "Es Publica",
            "default": true
          },
          "usuario_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usuario Id"
          }
        },
        "type": "object",
        "required": [
          "estudiante_id",
          "asignacion_id",
          "periodo_id",
          "texto"
        ],
        "title": "ObservacionResponse"
      },
...
      "PeriodoResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "numero": {
            "type": "integer",
            "title": "Numero"
          },
          "anio_id": {
            "type": "integer",
            "title": "Anio Id"
          },
          "activo": {
            "type": "boolean",
            "title": "Activo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "numero",
          "anio_id",
          "activo"
        ],
        "title": "PeriodoResponse"
      },
...
      "ResumenConvivenciaResponse": {
        "properties": {
          "estudiante_id": {
            "type": "integer",
            "title": "Estudiante Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "num_observaciones": {
            "type": "integer",
            "title": "Num Observaciones",
            "default": 0
          },
          "num_registros_negativos": {
            "type": "integer",
            "title": "Num Registros Negativos",
            "default": 0
          },
          "nota": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nota"
          },
          "nivel_nombre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nivel Nombre"
          },
          "supera_umbral": {
            "type": "boolean",
            "title": "Supera Umbral",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "estudiante_id",
          "nombre"
        ],
        "title": "ResumenConvivenciaResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}